Add option to mute auto-repeat sounds for prev/next button presses
This update adds an option to the Audio/Video dialog box that
allow you to mute the button sound for auto-repeat events.
When the prev/next or prevpage/nextpage buttons are pressed,
you will hear the normal button press sound, and if held, you
will hear one additional button press sound as the auto-repeat
scrolling kicks in.

This option effectively replicates the behaviour of PinballX when
the next/prev buttons are held down.  This also has the benefit
of slightly smoother scrolling as the wheel spins.