Add option to terminate a system via its process name
This commit adds an option to the System Dialog to allow a sytem
to be killed via process name.  This augments the existing options
to terminate by closing the window or terminate killing the process
by it's handle.

This was added to address an issue in Pinball FX3 where the game
will hang upon exit instead of closing gracefully if it was
originally launched with one of the "-hotseat_x" options.  The
issue was reported as issue #11 on github back in Oct. 2018 but
no resolution was identified at that time.

It's unclear why the built-in TerminateProcess() call is unable
to force Pinball FX3 to terminate, however, if I explicitly locate
a handle based on the process name, and then terminate that handle
via TerminateProcess(), the process immediatly terminates and returns
control to PinballY.